fix: tx details header
